Extract Server and IpcChannel
